feat: carried out fields validation into variables
