+++ src/components/ProfileForm/ProfileForm.module.css
@@ -5,5 +5,6 @@
 }
 
 .steps {
+  padding: 1rem 1rem 0rem;
   max-width: 900px;
 }
