+++ src/pages/Profile/Profile.tsx
@@ -2,7 +2,7 @@ import ProfilePage from '../../components/ProfileForm/ProfileForm';
 
 function Profile() {
   return (
-    <div className="wrapper">
+    <div className="wrapper" style={{ justifyContent: 'start' }}>
       <ProfilePage />
     </div>
   );
